Move `utils/firebaseConfig.ts` to `utils/firebase/firebaseConfig.ts`
Related to #56
